Admin. Sorting files 1.1
Admin. Sorting files 1.1
------------------------
What does the program do?
It shifts files to another folder according to the condition that the user assigns in the Enter a step value line field. For example, if you assign 2 to the Step field, then files will be transferred to an empty folder through two files - 1, 2, 3, 4, 5, 6, 7, 8, 9, 2, 5, 8 will be transferred.
----------------------------------
Order of actions:
Create the first folder, put the necessary files in the first folder, run the program, enter the necessary parameters in the field, click the Move Files button, select the first folder with the source files in the dialog box, then a second dialog box will appear to select the second folder that the program will create itself inside the first folder to move the files sorted by the program. 
+++++++++++++++++++++++++++++++++++
Что делает программа?
Она перекладывает файлы в другую папку по условию которое назначает пользователь в поле строки Enter a step value. Например если в поле Step назначить 2 то в пустую папку переложатся файлы через два файла - 1 2 3 4 5 6 7 8 9 будет переложено 2 5 8.
-----------------------------------
Порядок действий:
Создать первую папку, поместить в первую папку нужные файлы, запустить программу, ввести нужные параметры в поле, нажать кнопку Move Files, в диалоговом окне выбрать первую папку с исходными файлами, затем появится второе диалоговое окно для выбора второй папки которую программа создаст сама внутри первой папки для перемещения отсортированных программой файлов.